+++ scripts/differential-test/corpus/canary/pr/dromaeo-object-array-modern.js
@@ -1,5 +1,5 @@
 'use strict';
 
-require('../../../../../tests/performance/Benchmarks/Scenarios/dromaeo-object-array-modern.js');
+require('../../../../../tests/performance/Benchmarks/Scenarios/dromaeo/dromaeo-object-array-modern.js');
 
 console.log('completed=dromaeo-object-array-modern');